pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "guardduty-playbook-commons"
version = "0.1.0"
description = "Community-maintained GuardDuty finding remediation playbooks with multi-target converter"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
runbook = ["boto3>=1.28"]
dev = ["pytest>=7.0", "ruff>=0.4"]

[project.scripts]
gdpc = "converter.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["converter"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
